backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0

# WebSocket for real-time logs
Flask-SocketIO==5.3.6
eventlet==0.33.3

# Rate Limiting (Optional)
Flask-Limiter==3.5.0

# For bcrypt (used for password hashing)
Flask-Bcrypt==1.0.1
